Refactor: change RunMode enum to use uint8_t for reduced memory usage and update related functions

+++ yy_cybergear/include/yy_cybergear/protocol_types.hpp
@@ -81,7 +81,7 @@ constexpr uint16_t SPEED_KP = 0x701F;
 constexpr uint16_t SPEED_KI = 0x7020;
 
 // ======== Command & status types ========
-enum class RunMode : uint32_t { OperationControl = 0, Position = 1, Speed = 2, Current = 3 };
+enum class RunMode : uint8_t { OperationControl = 0, Position = 1, Speed = 2, Current = 3 };
 
 struct FaultWarning
 {
